+++ src/utils/render.js
@@ -1,0 +1,17 @@
+import {renderPositions} from "./util";
+
+export const render = (container, element, place) => {
+  switch (place) {
+    case renderPositions.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case renderPositions.BEFOREEND:
+      container.append(element);
+      break;
+  }
+};
+
+export const remove = (component) => {
+  component.getElement().remove();
+  component.removeElement();
+};
