+++ src/utils/util.js
@@ -36,18 +36,3 @@ export const createElement = (template) => {
   return newElement.firstChild;
 };
 
-export const render = (container, element, place) => {
-  switch (place) {
-    case renderPositions.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case renderPositions.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export const remove = (component) => {
-  component.getElement().remove();
-  component.removeElement();
-};
